OLA_Project/requirements.txt
streamlit
pandas
sqlalchemy
pymysql
plotly
